update the layout
